move `Footer` component out of `HomePage` dir
